refactor: rename features/presets/ → features/templates/
Folder name now matches the Template* file naming used since v0.9.0.
Underlying Supabase tables (presets, preset_actions) unchanged --
data/api/templates.js documents the table-name mapping inline.
