+++ src/admin-app/src/App.jsx
@@ -26,8 +26,8 @@ import DashboardPage from "./features/dashboard/DashboardPage";
 import GroupDetail from "./features/groups/GroupDetail";
 import GroupForm from "./features/groups/GroupForm";
 import GroupsPage from "./features/groups/GroupsPage";
-import TemplateForm from "./features/presets/TemplateForm";
-import TemplatesPage from "./features/presets/TemplatesPage";
+import TemplateForm from "./features/templates/TemplateForm";
+import TemplatesPage from "./features/templates/TemplatesPage";
 import ReportsPage from "./features/reports/ReportsPage";
 import UserDetail from "./features/users/UserDetail";
 import UserForm from "./features/users/UserForm";
