+++ src/admin-app/src/features/templates/components/TemplateFieldsSection.jsx
